+++ Point.class.cpp
@@ -8,11 +8,11 @@ Point::Point( void ) {
 
 }
 
-Point::Point( unsigned int const value,
-	unsigned int const x_current,
-	unsigned int const y_current,
-	unsigned int const x_final,
-	unsigned int const y_final ) :
+Point::Point( size_t const value,
+	size_t const x_current,
+	size_t const y_current,
+	size_t const x_final,
+	size_t const y_final ) :
 	value(value),
 	x_current(x_current),
 	y_current(y_current),
